requirements.txt
absl-py==2.0.0
backcall==0.2.0
beautifulsoup4==4.10.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
conda==4.10.3
conda-build==3.21.5
conda-package-handling==1.7.3
cryptography==3.4.8
cycler==0.11.0
decorator==5.1.0
dnspython==2.1.0
dotmap==1.3.30
filelock==3.0.12
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
glob2==0.7
google-auth==2.25.2
google-auth-oauthlib==0.4.6
grpcio==1.60.0
huggingface-hub==0.16.4
idna==2.10
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
inflect==6.0.5
ipython==7.27.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.4.5
libarchive-c==2.9
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.6.3
numpy==1.21.2
oauthlib==3.2.2
olefile==0.46
opencv-python-headless==4.8.1.78
packaging==23.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.0.1
pkginfo==1.7.1
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat==0.6.3
pycparser==2.20
pydantic==1.10.13
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2021.3
PyWavelets==1.3.0
PyYAML==5.4.1
regex==2023.12.25
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml-conda==0.15.100
scikit-image==0.19.3
scipy==1.7.3
setuptools==58.0.4
shapely==2.0.2
six==1.16.0
soupsieve==2.2.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
timm==0.6.12
torch==1.10.0
torchelastic==0.2.0
torchtext==0.11.0
torchvision==0.11.0
tqdm==4.61.2
traitlets==5.1.0
typing-extensions==4.7.1
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.2.3
wheel==0.36.2
zipp==3.15.0
